requirements.txt
grpcio
tensorflow
tensorflow-serving-api==1.10.1
requests
flask
